+++ mimesis/providers/address.py
@@ -34,7 +34,7 @@ class Address(BaseDataProvider):
         """
         super().__init__(*args, **kwargs)
         self._datafile = 'address.json'
-        self.pull(self._datafile)
+        self._pull(self._datafile)
 
     class Meta:
         """Class for metadata."""