+++ mimesis/providers/date.py
@@ -27,7 +27,7 @@ class Datetime(BaseDataProvider):
         """
         super().__init__(*args, **kwargs)
         self._datafile = 'datetime.json'
-        self.pull(self._datafile)
+        self._pull(self._datafile)
 
     class Meta:
         """Class for metadata."""